requirements.txt
pandas==1.1.3
streamlit==0.76.0
scikit-learn==0.23.2
matplotlib==3.3.2
Pillow~=9.0.1